+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Wildcard support for all entity lists (like the blight blacklist). Wildcard character is asterisk (*). Wildcards are only allowed at the end of the entry. Examples: `modid:*` or `modid:something_*`.
+- A separate pair of configs for adjusting absorption text position
 ### Changed
 - Damage scaling events now have highest priority, should improve compatibility with some mods